feat(examples): implement URL routing with dynamic routes for direct navigation
